moved command classes & commands around & did osme cleanup
