+++ src/main/java/de/kittybot/kittybot/commands/admin/KickCommand.java
@@ -28,18 +28,21 @@ public class KickCommand extends RunGuildCommand{
 			ia.error("I don't have the required permission to kick members");
 			return;
 		}
-		var userId = options.getLong("user");
-		ia.getGuild().retrieveMemberById(userId).queue(member -> {
-				if(!ia.getSelfMember().canInteract(member)){
-					ia.error("I can't interact with this member");
-					return;
-				}
-				var reason = options.getOrDefault("reason", "Kicked by " + ia.getMember().getAsMention());
-				ia.getGuild().kick(member, reason).reason(reason).queue(success ->
-						ia.reply("Kicked `" + MarkdownSanitizer.escape(member.getUser().getAsTag()) + "` with reason: " + reason),
-					error -> ia.error("Failed to kick " + MessageUtils.getUserMention(userId) + " for reason: `" + error.getMessage() + "`")
-				);
-			}, error -> ia.error("I could not find the provided user")
+		var user = options.getUser("user");
+		if(user.getIdLong() == ia.getUserId()){
+			ia.error("You can't kick yourself");
+		}
+		var member = options.getMember("user");
+		if(member != null){
+			if(!ia.getSelfMember().canInteract(member)){
+				ia.error("I can't interact with this member");
+				return;
+			}
+		}
+		var reason = options.getOrDefault("reason", "Kicked by " + ia.getMember().getAsMention());
+		ia.getGuild().kick(user.getId(), reason).reason(reason).queue(success ->
+				ia.reply("Kicked `" + MarkdownSanitizer.escape(user.getAsTag()) + "`(`" + user.getId() + "`).\nReason: " + reason),
+			error -> ia.error("Failed to kick " + user.getAsMention() + ".\nReason: `" + error.getMessage() + "`")
 		);
 	}
