--- src/main/java/de/kittybot/kittybot/objects/streams/youtube/YouTubeWrapper.java
@@ -1,5 +1,0 @@
-package de.kittybot.kittybot.objects.streams.youtube;
-
-public class YouTubeWrapper{
-	// TODO
-}
